semi correzione

+++ libft/libft.h
@@ -16,9 +16,9 @@
 
 typedef struct s_list
 {
-	void *content;
-	struct s_list *next;
-} t_list;
+	void				*content;
+	struct s_list		*next;
+}					t_list;
 
 char	**ft_split(char const *s, char c);
 int		ft_strlen(const char *s);
@@ -36,7 +36,7 @@ int		ft_strncmp(char *s1, char *s2, unsigned int n);
 char	*ft_strrchr(const char *s, int c);
 int		ft_tolower(int c);
 int		ft_toupper(int c);
-unsigned int	ft_strlcat(char *dest, char *src, unsigned int size);
+size_t	ft_strlcat(char *dest, char *src, unsigned int size);
 size_t	ft_strlcpy(char *dst, const char *src, size_t size);
 void	ft_bzero(void *s, size_t n);
 void	*ft_memcpy(void *dest, const void *src, size_t n);
@@ -47,19 +47,19 @@ int		ft_isalnum(int c);
 int		ft_isascii(int c);
 int		ft_atoi(const char *str);
 void	*ft_memchr(const void *s, int c, size_t n);
-char 	*ft_itoa(int n);
-void 	ft_striteri(char *s, void (*f)(unsigned int, char*));
-char 	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
+char	*ft_itoa(int n);
+void	ft_striteri(char *s, void (*f)(unsigned int, char*));
+char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 void	ft_putchar_fd(char c, int fd);
 void	ft_putendl_fd(char *s, int fd);
 void	ft_putnbr_fd(int n, int fd);
-void 	ft_putstr_fd(char *s, int fd);
-int 	ft_lstsize(t_list *lst);
+void	ft_putstr_fd(char *s, int fd);
+int		ft_lstsize(t_list *lst);
 t_list	*ft_lstnew(void *content);
 t_list	*ft_lstlast(t_list *lst);
-void 	ft_lstiter(t_list *lst, void (*f)(void *));
-void 	ft_lstdelone(t_list *lst, void (*del)(void *));
+void	ft_lstiter(t_list *lst, void (*f)(void *));
+void	ft_lstdelone(t_list *lst, void (*del)(void *));
 void	ft_lstclear(t_list **lst, void (*del)(void *));
-void ft_lstadd_front(t_list **lst, t_list *new);
-void ft_lstadd_back(t_list **lst, t_list *new);
+void	ft_lstadd_front(t_list **lst, t_list *new);
+void	ft_lstadd_back(t_list **lst, t_list *new);
 #endif
